requirements.txt
Django==4.0.1
python_bcrypt==0.3.2
